requirements.txt
time
RPi.GPIO
random
flask
